* currency amount type is changed to BigDecimal; * java documentation is added; * tests are corrected.

+++ src/main/java/com/crypto/recommendation/common/CryptoNormalizedRange.java
@@ -3,9 +3,11 @@ package com.crypto.recommendation.common;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 public class CryptoNormalizedRange {
     private String code;
-    private double normalizedRange;
+    private BigDecimal normalizedRange;
 }
